[준혁] 착지상태에서 이동 가능하게 변경
.

+++ FightingGame/Assets/Scripts/Network/Data/NetworkData.cs
@@ -10,5 +10,5 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "NetworkData", menuName = "Make NetworkData For Synchronize", order = 1)]
 public class NetworkData : ScriptableObject
 {
-	public int hpData = 0;
+	public float hpData = 0;
 }
